requirements.txt
# Minimal development requirements
openai>=1.0.0
anthropic>=0.18.0
pyyaml>=6.0
pytest>=7.0.0
pytest-asyncio>=0.21.0
